backend/requirements.txt
fastapi
uvicorn
ag2[openai,tinyfish]
python-dotenv
pydantic
sse-starlette
httpx
